pyproject.toml
[project]
name = "sec-log-analyzer-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "mkalitka",email = "mikolaj.kalitka@proton.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.10"

